+++ server/models/gaitDataPoints.js
@@ -3,13 +3,13 @@ module.exports = (sequelize, DataTypes) => {
   const gaitDataPoint = sequelize.define(
     "gaitDataPoint",
     {
-      ax: DataTypes.FLOAT(3, 2),
-      ay: DataTypes.FLOAT(3, 2),
-      az: DataTypes.FLOAT(3, 2),
-      gx: DataTypes.FLOAT(5, 2),
-      gy: DataTypes.FLOAT(5, 2),
-      gz: DataTypes.FLOAT(5, 2),
-      time: DataTypes.DOUBLE(20, 8)
+      ax: DataTypes.FLOAT(5),
+      ay: DataTypes.FLOAT(5),
+      az: DataTypes.FLOAT(5),
+      gx: DataTypes.FLOAT(7),
+      gy: DataTypes.FLOAT(7),
+      gz: DataTypes.FLOAT(7),
+      time: DataTypes.DOUBLE
     },
     {}
   );
